requirements.txt
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
packaging==24.1
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
sentence-transformers==3.2.0
service-identity==24.1.0
speedtest-cli==2.1.3
sympy==1.13.3
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.1.0
